Addressed review comments
* Fixed log message
* Moved AM-specific utils to pkg/admission
* Use constants in context.go
