+++ pkg/common/utils/utils.go
@@ -399,50 +399,3 @@ func GetTaskGroupsFromAnnotation(pod *v1.Pod) ([]v1alpha1.TaskGroup, error) {
 	}
 	return taskGroups, nil
 }
-
-// generate appID based on the namespace value,
-// and the max length of the ID is 63 chars.
-func generateAppID(namespace string) string {
-	generatedID := fmt.Sprintf("%s-%s-%s", constants.AutoGenAppPrefix, namespace, constants.AutoGenAppSuffix)
-	appID := fmt.Sprintf("%.63s", generatedID)
-	return appID
-}
-
-func UpdatePodLabelForAdmissionController(pod *v1.Pod, namespace string) map[string]string {
-	existingLabels := pod.Labels
-	result := make(map[string]string)
-	for k, v := range existingLabels {
-		result[k] = v
-	}
-
-	sparkAppID := GetPodLabelValue(pod, constants.SparkLabelAppID)
-	appID := GetPodLabelValue(pod, constants.LabelApplicationID)
-	if sparkAppID == "" && appID == "" {
-		// if app id not exist, generate one
-		// for each namespace, we group unnamed pods to one single app
-		// application ID convention: ${AUTO_GEN_PREFIX}-${NAMESPACE}-${AUTO_GEN_SUFFIX}
-		generatedID := generateAppID(namespace)
-		result[constants.LabelApplicationID] = generatedID
-
-		// if we generate an app ID, disable state-aware scheduling for this app
-		if _, ok := existingLabels[constants.LabelDisableStateAware]; !ok {
-			result[constants.LabelDisableStateAware] = "true"
-		}
-	}
-
-	if _, ok := existingLabels[constants.LabelQueueName]; !ok {
-		result[constants.LabelQueueName] = constants.DefaultQueue
-	}
-
-	return result
-}
-
-func UpdatePodAnnotationForAdmissionController(pod *v1.Pod, key string, value string) map[string]string {
-	existingAnnotations := pod.Annotations
-	result := make(map[string]string)
-	for k, v := range existingAnnotations {
-		result[k] = v
-	}
-	result[key] = value
-	return result
-}
