+++ pkg/common/utils/utils_test.go
@@ -20,7 +20,6 @@ package utils
 
 import (
 	"fmt"
-	"strings"
 	"testing"
 	"time"
 
@@ -1019,161 +1018,6 @@ func TestGetTaskGroupFromAnnotation(t *testing.T) {
 	assert.Equal(t, taskGroups2[0].MinMember, int32(3))
 	assert.Equal(t, taskGroups2[0].MinResource["cpu"], resource.MustParse("2"))
 	assert.Equal(t, taskGroups2[0].MinResource["memory"], resource.MustParse("1Gi"))
-}
-
-// nolint: funlen
-func TestUpdatePodLabelForAdmissionController(t *testing.T) {
-	// verify when appId/queue are not given,
-	pod := &v1.Pod{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Pod",
-			APIVersion: "v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:            "a-test-pod",
-			Namespace:       "default",
-			UID:             "7f5fd6c5d5",
-			ResourceVersion: "10654",
-			Labels: map[string]string{
-				"random": "random",
-			},
-		},
-		Spec:   v1.PodSpec{},
-		Status: v1.PodStatus{},
-	}
-
-	if result := UpdatePodLabelForAdmissionController(pod, "default"); result != nil {
-		assert.Equal(t, len(result), 4)
-		assert.Equal(t, result["random"], "random")
-		assert.Equal(t, result["queue"], "root.default")
-		assert.Equal(t, result["disableStateAware"], "true")
-		assert.Equal(t, strings.HasPrefix(result["applicationId"], constants.AutoGenAppPrefix), true)
-	} else {
-		t.Fatal("UpdatePodLabelForAdmissionController is not as expected")
-	}
-
-	// verify if applicationId is given in the labels,
-	// we won't modify it
-	pod = &v1.Pod{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Pod",
-			APIVersion: "v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:            "a-test-pod",
-			Namespace:       "default",
-			UID:             "7f5fd6c5d5",
-			ResourceVersion: "10654",
-			Labels: map[string]string{
-				"random":        "random",
-				"applicationId": "app-0001",
-			},
-		},
-		Spec:   v1.PodSpec{},
-		Status: v1.PodStatus{},
-	}
-
-	if result := UpdatePodLabelForAdmissionController(pod, "default"); result != nil {
-		assert.Equal(t, len(result), 3)
-		assert.Equal(t, result["random"], "random")
-		assert.Equal(t, result["queue"], "root.default")
-		assert.Equal(t, result["applicationId"], "app-0001")
-	} else {
-		t.Fatal("UpdatePodLabelForAdmissionController is not as expected")
-	}
-
-	// verify if queue is given in the labels,
-	// we won't modify it
-	pod = &v1.Pod{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Pod",
-			APIVersion: "v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:            "a-test-pod",
-			Namespace:       "default",
-			UID:             "7f5fd6c5d5",
-			ResourceVersion: "10654",
-			Labels: map[string]string{
-				"random": "random",
-				"queue":  "root.abc",
-			},
-		},
-		Spec:   v1.PodSpec{},
-		Status: v1.PodStatus{},
-	}
-	if result := UpdatePodLabelForAdmissionController(pod, "default"); result != nil {
-		assert.Equal(t, len(result), 4)
-		assert.Equal(t, result["random"], "random")
-		assert.Equal(t, result["queue"], "root.abc")
-		assert.Equal(t, result["disableStateAware"], "true")
-		assert.Equal(t, strings.HasPrefix(result["applicationId"], constants.AutoGenAppPrefix), true)
-	} else {
-		t.Fatal("UpdatePodLabelForAdmissionControllert is not as expected")
-	}
-
-	// namespace might be empty
-	// labels might be empty
-	pod = &v1.Pod{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Pod",
-			APIVersion: "v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:            "a-test-pod",
-			UID:             "7f5fd6c5d5",
-			ResourceVersion: "10654",
-		},
-		Spec:   v1.PodSpec{},
-		Status: v1.PodStatus{},
-	}
-	if result := UpdatePodLabelForAdmissionController(pod, "default"); result != nil {
-		assert.Equal(t, len(result), 3)
-		assert.Equal(t, result["queue"], "root.default")
-		assert.Equal(t, result["disableStateAware"], "true")
-		assert.Equal(t, strings.HasPrefix(result["applicationId"], constants.AutoGenAppPrefix), true)
-	} else {
-		t.Fatal("UpdatePodLabelForAdmissionController is not as expected")
-	}
-
-	// pod name might be empty, it can comes from generatedName
-	pod = &v1.Pod{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Pod",
-			APIVersion: "v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			GenerateName: "some-pod-",
-		},
-		Spec:   v1.PodSpec{},
-		Status: v1.PodStatus{},
-	}
-	if result := UpdatePodLabelForAdmissionController(pod, "default"); result != nil {
-		assert.Equal(t, len(result), 3)
-		assert.Equal(t, result["queue"], "root.default")
-		assert.Equal(t, result["disableStateAware"], "true")
-		assert.Equal(t, strings.HasPrefix(result["applicationId"], constants.AutoGenAppPrefix), true)
-	} else {
-		t.Fatal("UpdatePodLabelForAdmissionController is not as expected")
-	}
-
-	pod = &v1.Pod{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Pod",
-			APIVersion: "v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{},
-		Spec:       v1.PodSpec{},
-		Status:     v1.PodStatus{},
-	}
-	if result := UpdatePodLabelForAdmissionController(pod, "default"); result != nil {
-		assert.Equal(t, len(result), 3)
-		assert.Equal(t, result["queue"], "root.default")
-		assert.Equal(t, result["disableStateAware"], "true")
-		assert.Equal(t, strings.HasPrefix(result["applicationId"], constants.AutoGenAppPrefix), true)
-	} else {
-		t.Fatal("UpdatePodLabelForAdmissionController is not as expected")
-	}
 }
 
 func TestGetCoreSchedulerConfigFromConfigMapNil(t *testing.T) {
